+++ src/buffer.h
@@ -1,8 +1,9 @@
 #ifndef __BUFFER_H__
 #define __BUFFER_H__
 
-#include <stdio.h>
+#include <stddef.h>
+#include <stdint.h>
 
-void** adjust_buffer (void**, size_t const, size_t const);
+void** adjust_buffer (void**, uint64_t const, uint64_t const);
 
 #endif /* __BUFFER_H__ */
