+++ src/queue.h
@@ -14,7 +14,7 @@ void* remove_head_of_queue (fifo_t *const queue);
 void* peek_tail_of_queue (fifo_t const*const queue);
 void* peek_head_of_queue (fifo_t const*const queue);
 
-void* get_queue_element (fifo_t const*const queue, size_t const position);
+void* get_queue_element (fifo_t const*const queue, uint64_t const position);
 
 void insert_at_tail_of_queue (fifo_t *const queue, void *const element);
 void insert_at_head_of_queue (fifo_t *const queue, void *const element);
